requirement.txt
streamlit
python-dotenv
pypdf
pinecone-client
google-genai
openai
tiktoken

requests
uuid
